Completed phase 1

+++ backend.py
@@ -17,6 +17,18 @@ def multi(d):
 	return d['x']*d['y']
 
 ########
+def Is_login(session):
+  if "login_key" in session and "id" in session["login_key"]:
+    return True
+  else:
+    return False
+
+
+def login_id(session):
+	if Is_login(session):
+		return session["login_key"]["id"]
+	else:
+		return 0
 
 @backend.api('/sign_up')
 def sign_up_function(frontend_dict, session):
@@ -26,7 +38,7 @@ def sign_up_function(frontend_dict, session):
 		query2 = "insert into users (name, email, phone, password) values ({name}, {email}, {phone}, {password}) "
 		db.writeQuery(query2, frontend_dict)
 		new_id = db.readQuery("select max(id) as new_id from users")[0]["new_id"]
-		session['login'] = {"id": new_id, "name": frontend_dict["name"]}
+		session['login_key'] = {"id": new_id, "name": frontend_dict["name"]}
 	else:
 		output['error']="phle se email h"
 	return output
@@ -36,7 +48,7 @@ def sign_up_function(frontend_dict, session):
 def login_function(frontend_dict, session):
 	query_output = db.readQuery("select * from users where email={email} AND password={password}", frontend_dict)
 	if len(query_output) != 0:
-		session['login'] = {"id": query_output[0]['id'], "name": query_output[0]['name']}
+		session['login_key'] = {"id": query_output[0]['id'], "name": query_output[0]['name']}
 		return {}
 	else:
 		return {"error": "Email or password is incorrect."}
@@ -52,9 +64,10 @@ def save_products(frontend_dict):
 	return db.readQuery("select * from products order by id desc")
 
 @backend.api('/all_products')
-def all_products_function(frontend_dict):
-	query="select products.*, COALESCE(cart.product_quantity, 0) as product_quantity from products left join cart  on products.id=cart.product_id order by products.id desc "
-	return db.readQuery( query )
+def all_products_function(frontend_dict,session):
+	frontend_dict["user_id"] = login_id(session)
+	query="select products.*, COALESCE(cart.product_quantity, 0) as product_quantity from products left join cart  on products.id=cart.product_id AND cart.user_id={user_id} order by products.id desc "
+	return db.readQuery( query,frontend_dict)
 
 @backend.api('/product_update')
 def update(frontend_dict):
@@ -65,30 +78,30 @@ def update(frontend_dict):
 
 
 @backend.api('/add_product_quantity')
-def add_product_quantity(frontend_dict):
-	db.writeQuery("delete from cart where  product_id={product_id}", frontend_dict)
+def add_product_quantity(frontend_dict, session):
+	print(session)
+	frontend_dict["user_id"] = session["login_key"]["id"]
+	db.writeQuery("delete from cart where product_id={product_id} AND user_id = {user_id}",frontend_dict)
 	if frontend_dict["product_quantity"] > 0:
-		query = "insert into cart (product_id ,product_quantity) values ({product_id}, {product_quantity}) "
+		query = "insert into cart (product_id ,product_quantity,user_id) values ({product_id}, {product_quantity},{user_id})"
 		db.writeQuery(query, frontend_dict)
 
 @backend.api('/products_in_cart')
-def products_in_cart(frontend_dict):
-	query1=" select products.id, products.name, products.price, products.discount,products.image, cart.product_quantity , cart.product_quantity * products.price as subtotal from cart inner join products on cart.product_id=products.id "
-	query2="select sum(cart.product_quantity * products.price) as total_amount from cart inner join products on cart.product_id=products.id "
-	query2_output = db.readQuery(query2)
+def products_in_cart(frontend_dict, session):
+	frontend_dict["user_id"] = session["login_key"]["id"]
+	query1=" select products.id, products.name, products.price, products.discount,products.image, cart.product_quantity , cart.product_quantity * products.price as subtotal from cart inner join products on cart.product_id=products.id where cart.user_id = {user_id}"
+	query2="select sum(cart.product_quantity * products.price) as total_amount from cart inner join products on cart.product_id=products.id where cart.user_id = {user_id}"
+	query2_output = db.readQuery(query2 , frontend_dict)
 	total_amount = query2_output[0]["total_amount"]
 	p = {
-		"total_list": db.readQuery(query1),
+		"total_list": db.readQuery(query1, frontend_dict),
 		"total_amount": total_amount}
 	return p
 
 @backend.api('/add_product_quantity_in_cart')
-def add_product_quantity_in_cart(frontend_dict):
-	db.writeQuery("delete from cart where  product_id={product_id}", frontend_dict)
-	if frontend_dict["product_quantity"] > 0:
-		query = "insert into cart (product_id ,product_quantity) values ({product_id}, {product_quantity}) "
-		db.writeQuery(query, frontend_dict)
-	return products_in_cart(frontend_dict)
+def add_product_quantity_in_cart(frontend_dict,session):
+	add_product_quantity(frontend_dict,session)
+	return products_in_cart(frontend_dict, session)
 
 
 @backend.api('/products_according_to_id')
@@ -101,12 +114,13 @@ def products_according_to_id(frontend_dict):
 		return output[0]
 
 @backend.api('/checkout')
-def checkout(frontend_dict):
-	cart_info=products_in_cart({});
-	query1="insert into orders(store_name , total_amount, ordered_at, expected_delivery_at, status) values ('sonu store', {total_amount}, {ordered_at}, {expected_delivery_at}, 'ORDERED')"
+def checkout(frontend_dict, session):
+	cart_info=products_in_cart({}, session);
+	query1="insert into orders(store_name , total_amount, ordered_at, expected_delivery_at, status, user_id) values ('sonu store', {total_amount}, {ordered_at}, {expected_delivery_at}, 'ORDERED', {user_id})"
 	frontend_dict['total_amount'] = cart_info['total_amount']
 	frontend_dict['ordered_at'] = int(time.time())
 	frontend_dict['expected_delivery_at'] = frontend_dict['ordered_at'] + 48*3600
+	frontend_dict["user_id"] = session["login_key"]["id"]
 	db.writeQuery(query1, frontend_dict)
 
 	A="select max(id) As order_id from orders"
@@ -119,8 +133,10 @@ def checkout(frontend_dict):
 	db.writeQuery("delete from cart")
 
 @backend.api('/orders')
-def orders(frontend_dict):
-	l=db.readQuery("select * from orders order by id desc")
+def orders(frontend_dict,session):
+	frontend_dict["user_id"] = session["login_key"]["id"]
+	query1="select * from orders where orders.user_id = {user_id} order by id desc"
+	l=db.readQuery( query1, frontend_dict)
 	for i in l:
 		i['ordered_at_str'] = time.strftime('%Y-%m-%d %I:%M:%S %p', time.localtime(i['ordered_at']))
 		i['delivery_at_str'] = time.strftime('%Y-%m-%d %I:%M:%S %p', time.localtime(i['delivery_at']))
@@ -144,6 +160,10 @@ def orders_details(frontend_dict):
 def Status_save(frontend_dict):
 	db.writeQuery("UPDATE orders SET status = {status} WHERE id={order_id}", frontend_dict)
 	return db.readQuery("select * from orders order by id desc")
+
+
+
+  
 
 
 backend.run(port=5502)
